+++ src/apps/oauth-client/main.ts
@@ -4,10 +4,10 @@ import process, {
 
 import {
   config,
-} from '#@/src/Contexts/OauthClient/Shared/infrastructure/Config/config.js'
+} from '#/src/Contexts/OauthClient/Shared/infrastructure/Config/config.js'
 import {
   error,
-} from '#@/src/Contexts/Shared/infrastructure/Logger/PinoLogger.js'
+} from '#/src/Contexts/Shared/infrastructure/Logger/PinoLogger.js'
 
 import {
   AppBackend,