+++ tests/apps/oauth-client/features/step_definitions/hooks.steps.ts
@@ -15,13 +15,13 @@ import type TestAgent from 'supertest/lib/agent.js'
 
 import {
   AppBackend,
-} from '#@/src/apps/oauth-client/server.js'
+} from '#/src/apps/oauth-client/server.js'
 import {
   config,
-} from '#@/src/Contexts/OauthClient/Shared/infrastructure/Config/config.js'
+} from '#/src/Contexts/OauthClient/Shared/infrastructure/Config/config.js'
 import {
   ONE_SECOND_IN_MILLISECONDS,
-} from '#@/src/Contexts/Shared/domain/time.js'
+} from '#/src/Contexts/Shared/domain/time.js'
 
 const application = new AppBackend(config)